added scroll placeholder when scrolling too fast and broke out components
