+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.0",
   "scripts": {
     "dev": "vite",
-    "build": " vite build",
+    "build": "vite build",
     "preview": "vite preview",
     "generate": "graphql-codegen"
   },
@@ -19,10 +19,6 @@
     "@graphql-codegen/typescript-react-apollo": "^3.2.16",
     "@graphql-codegen/typescript-react-query": "^3.5.14",
     "@headlessui/react": "^1.7.3",
-    "@mui/base": "^5.0.0-alpha.87",
-    "@mui/material": "^5.8.6",
-    "@tanstack/react-table": "^8.5.15",
-    "@tanstack/react-virtual": "^3.0.0-beta.18",
     "@types/react-window": "^1.8.5",
     "daisyui": "^2.31.0",
     "graphql": "^16.5.0",
@@ -34,13 +30,13 @@
     "react-hotkeys-hook": "^3.4.7",
     "react-query": "^3.39.1",
     "react-router-dom": "^6.3.0",
+    "react-split": "^2.0.14",
     "react-toastify": "^9.0.8",
     "react-virtuoso": "^2.19.1",
     "react-window": "^1.8.7",
     "vite-tsconfig-paths": "^3.5.0"
   },
   "devDependencies": {
-    "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/typography": "^0.5.7",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
