+++ src/component/HeroLoading.tsx
@@ -1,5 +1,5 @@
 export const HeroLoading = () => (
-    <div className="hero min-h-screen">
+    <div className="hero h-max">
         <div className="hero-content flex-col lg:flex-row">
             <div className="absolute right-1/2 bottom-1/2  transform translate-x-1/2 translate-y-1/2 ">
                 <div className="border-t-transparent border-solid animate-spin  rounded-full border-primary border-8 h-64 w-64"></div>
