+++ src/component/detailedReview/commentForm/CommentFormModalWrapper.tsx
@@ -1,0 +1,19 @@
+import { useAtomValue } from "jotai"
+import { useMemo } from "react"
+import { openCommentModalAtom } from "state/Atoms"
+import { CommentFormModal } from "./CommentFormModal"
+
+export const CommentFormModalWrapper = () => {
+    const modalData = useAtomValue(openCommentModalAtom)
+    const open = useMemo(() => modalData !== undefined, [modalData])
+
+    return (
+        <CommentFormModal
+            open={open}
+            title={modalData?.title ?? ""}
+            onCancel={modalData?.onCancel ?? (() => { })}
+            onSubmit={modalData?.onSubmit ?? (() => Promise.resolve())}
+            initialValue={modalData?.initialValue}
+        />
+    )
+}
