+++ tailwind.config.js
@@ -7,5 +7,5 @@ module.exports = {
   theme: {
     extend: {},
   },
-  plugins: [require("@tailwindcss/typography"), require('@tailwindcss/forms'), require("daisyui")],
+  plugins: [require("@tailwindcss/typography"), require("daisyui")],
 }
